pyproject.toml
[project]
name = "chat_bot"
version = "0.1.0"
description = ""
authors = [
    { name = "morteza", email = "khodajou.morteza@gmail.com" }
]
readme = "README.md"
requires-python = ">=3.12, <4.0"
dependencies = [
    "ruff (>=0.11.2,<0.12.0)",
    "ipython (>=9.0.2,<10.0.0)",
    "notebook (>=7.3.3,<8.0.0)",
    "azure-ai-documentintelligence (>=1.0.2,<2.0.0)",
    "pydantic (>=2.11.0,<3.0.0)",
    "pydantic-settings (>=2.8.1,<3.0.0)",
    "pandas (>=2.2.3,<3.0.0)",
    "tqdm (>=4.67.1,<5.0.0)",
    "pyarrow (>=19.0.1,<20.0.0)",
    "fastparquet (>=2024.11.0,<2025.0.0)",
    "mcp (>=1.6.0,<2.0.0)",
    "openai (>=1.71.0,<2.0.0)",
    "fastapi (>=0.115.12,<0.116.0)",
    "chromadb (>=0.4.24,<0.5.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 180
# other rules:
# * "DJ" for Django
# * "PYI" for type stubs etc.
# * "PD" for PandasVet
# * "NPY" for NumPy

lint.select = [
    "E",
    "W", # PyCodeStyle
    "F", # PyFlakes
    "C90", # McCabe (Function Complexity
    "I", # ISort
    "N", # PEP8 Naming Conventions
    "D", # PyDocStyle
    "UP", # pyupgrade
    "YTT", # flake8-2020 ('Yield from' etc.)
    "ANN", # flake8-annotations (missing type annotations)
    "ASYNC", # flake8-async (various async issues)
    "S", # flake8-bandit (security issues)
    "BLE", # blind exceptions
    "FBT", # boolean traps (Anti-Pattern, Google it.)
    "B", # bugbear (various anti-patterns)
    "A", # flake8-builtins (shadowing builtins)
    "COM", # flake8-commas (comma placement at line-end)
    "C4", # flake8-comprehensions (comprehension issues like unnecessary list comprehensions etc.)
    "DTZ", # avoid usage of naive datetime objects
    "T10", # watch for Debugger imports
    "EM", # ensure error messages are not formatted as f-strings and similar
    "FA", # flake8-future-annotations (ensure type hint annotation use mnodern syntax)
    "ISC", # implicit string concatenation
    "G", # flake8-logging-format (ensure logging format strings are valid)
    "INP", # do not use implicit namspace packages
    "PIE", # various anti-patterns and misfeatures
    "T20", # watch for print() calls
    "PT", # pytest style issues
    "Q", # quotes (ensure consistent usage of single/double quotes)
    "RSE", # some "raise" syntax issues
    "RET", # return values anti-patterns
    "SLF", # flake8-self (do not access "_private" attributes from outside)
    "SLOT", # flake8-slots (ensure usage of __slots__)
    "SIM", # flake8-simplify (simplify various constructs)
    "INT", # gettext issues (format strings etc.)
    "ARG", # disallow unused arguments
    "PTH", # use pathlib instead of os.path
    "TD", # enforce some syntax on TODO comments
    "FIX", # highlight TODO, FIXME, XXX etc.
    "PGH", # pygrep-hooks (policing "noqa" and similar)
    "PL", # PyLint (various issues)
    "TRY", # try/except/else/finally anti-patterns (try.ceratops)
    "FLY", # join vs. f-strings
    "PERF", # various performance issues
    "FURB", # modernize various constructs
    "LOG", # logging issues
    "RUF", # ruff (various issues)
]

lint.ignore = [
    "ANN204", # do not require "__init__" return type annotations
    "COM812", # allow comma placement at line-end
    "ISC001", # allow implicit string concatenation
    "D104", # do not require docstrings for public methods
    "D107", # do not require docstrings for public classes and modules
    "DTZ007", # no need for %z in strftime
    "RUF013", # implicit Optional type
    "FA102", # no from __future__ import annotations needed
    "ANN002", # no type annotation for *args
    "ANN003", # no type annotation for **kwargs
    "ARG002",
    "ARG003",
    "D100",
    "D101",
    "D102",
    "D103",
]

[tool.ruff.lint.pydocstyle]
convention = "google"
